Framework for logging implemented
Added framework for logging to allow for only unique detections to be logged. Threshold set to 5s (5000ms) in Display.h to prevent cluttering in case someone wants to see a specific time at which a person was spotted on a specific camera for video review purposes in security applications. This is just a framework update in preparation for logs, actual log writing will be implemented in next release.